+++ lib/hooks/useScene/__tests__/useScene.test.ts
@@ -19,7 +19,7 @@ fdescribe("useScene", () => {
 
     const expectDefaultScene: IScene = {
       id: expect.anything(),
-      name: "",
+      name: "Scene Name",
       group: undefined,
       aspects: {},
       gm: {
@@ -70,7 +70,7 @@ fdescribe("useScene", () => {
       });
 
       // THEN
-      expect(result.current.state.scene.name).toEqual("");
+      expect(result.current.state.scene.name).toEqual("Scene Name");
       expect(result.current.state.dirty).toEqual(false);
 
       // GIVEN
@@ -686,7 +686,7 @@ fdescribe("useScene", () => {
       act(() => {
         result.current.actions.resetScene();
       });
-      expect(result.current.state.scene.name).toEqual("");
+      expect(result.current.state.scene.name).toEqual("Scene Name");
       expect(Object.keys(result.current.state.scene.aspects).length).toEqual(0);
       expect(result.current.state.scene.players).toEqual([
         {
@@ -731,7 +731,7 @@ fdescribe("useScene", () => {
       act(() => {
         result.current.actions.resetScene();
       });
-      expect(result.current.state.scene.name).toEqual("");
+      expect(result.current.state.scene.name).toEqual("Scene Name");
       expect(Object.keys(result.current.state.scene.aspects).length).toEqual(1);
     });
   });